requirements.txt
azure-identity
azure-kusto-data
pyodbc
